Moves Set from space to data
